split k8sobjects
